add nodebin to imagescraper.sh, now works on different distros

+++ imagescraper.sh
@@ -1,27 +1,38 @@
 #!/bin/sh
-
+#
+# imagescraper.sh
+# Scrapes images from images-index.html
+#
 # Input arguments:
 # $1 - html source folder
 # $2 - URL to scrape
 # $3 - Image folder
+# $4 - File/path to node executable
+#
+# H.Dahle, 2020
 
-SOURCE=$1
+SRCFOLDER=$1
 JSFILE="cb.js"
 IMGFOLDER=$2
 URL=$3
+NODEBIN=$4
 DEST=$(mktemp -d -t scrape-XXXXXXXX)
 PORT=8081
 
-if [ "$SOURCE" = "" ] ; then
-  echo "Usage: $0 <source folder> <output folder>  [url]"
+if [ "$NODEBIN" = "" ] ; then
+  echo "Usage: $0 <source folder> <output folder> <url> <nodebin>"
+  echo "  <source folder>  Source HTML  folder"
+  echo "  <output folder>  Where to write the PNG files"
+  echo "  <url>            URL to scrape"
+  echo "  <nodebin>        Full path to node executable"
   echo "Example:"
   echo "  $0 ~/dashboard img"
-  echo "  $0 ~/dashboard img http://localhost:8080"
+  echo "  $0 ~/dashboard img http://localhost"
   exit
 fi
 
-if [ ! -d "${SOURCE}" ] ; then
-  echo "Source directory not found: " ${SOURCE} 
+if [ ! -d "${SRCFOLDER}" ] ; then
+  echo "Source directory not found: " ${SRCFOLDER} 
   exit
 fi
 
@@ -38,16 +49,27 @@ if [ ! -d "${IMGFOLDER}" ] ; then
   exit
 fi
 
+if [ "${URL}" = "" ] ; then
+  echo "URL to scrape is missing"
+  exit
+fi
+
+if [ ! -f "${NODEBIN}" ] ; then
+  echo "Node.js executable not found at: ${NODEBIN}"
+  exit
+fi
+
 # Copy files from HTML folder to local folder
-echo "Copyng files from ${SOURCE} to ${DEST}"
+echo "Copyng files from ${SRCFOLDER} to ${DEST}"
 rmdir ${DEST}
-cp -r ${SOURCE} ${DEST}
+cp -r ${SRCFOLDER} ${DEST}
 
 # Copy the CB.JS file containing callback code for PNG generation
 cp ${JSFILE} ${DEST}/js
 
 # Run web-server in the background, exit after 30 sec
-node server.js --folder ${DEST} --selfdestruct --port ${PORT} &
+${NODEBIN} server.js --folder ${DEST} --selfdestruct --port ${PORT} &
 
 # Scrape the images
-node scraper.js --url ${URL} --folder ${IMGFOLDER}
+echo "Mustfix: use of PORT is inconsistent, ugly and just bad"
+${NODEBIN} scraper.js --url ${URL}:${PORT} --folder ${IMGFOLDER}
